backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.4
django-cors-headers>=4.3,<4.4
psycopg2-binary>=2.9,<2.10
celery>=5.3,<5.4
redis>=5.0,<5.1
gunicorn>=21.2,<21.3
python-dotenv>=1.0,<1.1
django-filter>=24.0,<25.0
whitenoise>=6.6,<6.7
dj-database-url>=2.1,<2.2
